Remove reply queue
This commit removes the use of a different queue for reply messages.
The way that we implement the RPC system do not required the creation
of a different queue to receive reply messages.
We just bind a routing key on an existing queue and pass this routing
key on the "reply to" property of the sent message.